+++ run.py
@@ -1,0 +1,3 @@
+from discus.api.run import app
+
+app.run(debug=True, port=8080)
